+++ colors.js
@@ -12,6 +12,24 @@ module.exports = {
     75: 'rgb(var(--color-gold) / 0.75)',
     100: 'rgb(var(--color-gold) / 1)',
   },
+  'hot-pink': {
+    25: 'rgb(var(--color-hot-pink) / 0.25)',
+    50: 'rgb(var(--color-hot-pink) / 0.50)',
+    75: 'rgb(var(--color-hot-pink) / 0.75)',
+    100: 'rgb(var(--color-hot-pink) / 100)',
+  },
   primary: 'rgb(var(--text-primary) / 1)',
   'backg-primary': 'rgb(var(--backg-primary) / 1)',
+  'black-alpha': {
+    25: 'rgb(var(--black-100) / 0.25)',
+    50: 'rgb(var(--black-100) / 0.50)',
+    75: 'rgb(var(--black-100) / 0.75)',
+    100: 'rgb(var(--black-100) / 100)',
+  },
+  'white-alpha': {
+    25: 'rgb(var(--white-100) / 0.25)',
+    50: 'rgb(var(--white-100) / 0.50)',
+    75: 'rgb(var(--white-100) / 0.75)',
+    100: 'rgb(var(--white-100) / 100)',
+  },
 };